move folders and rename
