--- lab3/a.txt
@@ -1,1 +1,0 @@
-aaaaa
